fastapi-backend/requirements.txt
fastapi
uvicorn
langgraph
langchain-core
langchain-openai
psycopg_pool
pydantic
SQLAlchemy
psycopg[binary]
psycopg-pool
langgraph-checkpoint-postgres
python-dotenv